chore: add automatic module name
Maven Java Compiler places JARs will place JARs with an automatic
module name in a manifest on a module path instead of a classpath
so these JARs can be consumed by other modules.

An alternative would be to provide a full-blown module-info.java,
but that would make the JAR JDK9+ only. An automatic module name
is a good compromise: The JAR can target JDK8 and it still has
an explicit module name, so Maven treats it as a module instead of
a JAR on a classpath.